requirements.txt
adal==1.2.7
ansible==4.4.0
ansible-core==2.11.3
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.10
argcomplete==1.12.3
azure-appconfiguration==1.1.1
azure-batch==11.0.0
azure-cli==2.27.1
azure-cli-core==2.27.1
azure-cli-telemetry==1.0.6
azure-common==1.1.27
azure-core==1.17.0
azure-cosmos==3.2.0
azure-datalake-store==0.0.52
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-identity==1.6.0
azure-keyvault==1.1.0
azure-keyvault-administration==4.0.0b3
azure-loganalytics==0.1.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==0.2.0
azure-mgmt-appconfiguration==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==0.61.0
azure-mgmt-batch==16.0.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==0.3.0
azure-mgmt-cdn==11.0.0
azure-mgmt-cognitiveservices==12.0.0
azure-mgmt-compute==22.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==8.1.0
azure-mgmt-containerservice==16.0.0
azure-mgmt-core==1.2.2
azure-mgmt-cosmosdb==6.4.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==9.0.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==9.0.0
azure-mgmt-eventhub==4.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==8.0.0
azure-mgmt-imagebuilder==0.4.0
azure-mgmt-iotcentral==9.0.0b1
azure-mgmt-iothub==2.0.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==9.0.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==11.0.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==3.1.0
azure-mgmt-monitor==2.0.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==4.0.0
azure-mgmt-network==19.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.0.0
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==9.1.0b1
azure-mgmt-recoveryservices==2.0.0
azure-mgmt-recoveryservicesbackup==0.14.0
azure-mgmt-redhatopenshift==0.1.0
azure-mgmt-redis==13.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==18.0.0
azure-mgmt-search==8.0.0
azure-mgmt-security==0.6.0
azure-mgmt-servicebus==6.0.0
azure-mgmt-servicefabric==0.5.0
azure-mgmt-servicefabricmanagedclusters==1.0.0
azure-mgmt-signalr==1.0.0b2
azure-mgmt-sql==3.0.1
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==18.0.0
azure-mgmt-synapse==2.0.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==2.0.0
azure-multiapi-storage==0.6.2
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.6.0
azure-synapse-spark==0.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.2
Deprecated==1.2.12
fabric==2.6.0
humanfriendly==9.2
idna==2.10
invoke==1.6.0
isodate==0.6.0
javaproperties==0.5.1
Jinja2==3.0.1
jmespath==0.10.0
jsmin==2.2.2
jsondiff==1.2.0
knack==0.8.2
MarkupSafe==2.0.1
msal==1.13.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
ntlm-auth==1.5.0
oauthlib==3.1.1
packaging==20.9
paramiko==2.7.2
pathlib2==2.3.6
pkginfo==1.7.1
portalocker==1.7.1
psutil==5.8.0
pycparser==2.20
PyGithub==1.55
Pygments==2.9.0
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2019.1
pywinrm==0.4.2
PyYAML==5.4.1
requests==2.25.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
resolvelib==0.5.4
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scp==0.13.6
semver==2.13.0
six==1.16.0
sshtunnel==0.1.5
tabulate==0.8.9
urllib3==1.26.6
vsts==0.1.25
websocket-client==0.56.0
winrm==0.4.1
wrapt==1.12.1
xmltodict==0.12.0
